refactor: extract duplicated rule helpers into shared common module
Move get_source_line, walk_tree, make_finding, and make_finding_from_offsets
from 11 rule files into src/rules/common.rs to eliminate ~520 lines of
copy-pasted code. Add bounds checking to get_source_line so it returns an
empty string instead of panicking on out-of-range offsets.
